Feat: 유저가 작성한 스레드 목록 페이지네이션으로 가져오는 API 구현

+++ src/main/java/threads/server/domain/comment/repository/CommentRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import threads.server.domain.comment.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
-    Integer countByPostId(Long postId);
+    Long countByPostId(Long postId);
 }
