+++ src/main/java/threads/server/domain/like/repository/LikePostRepository.java
@@ -6,7 +6,7 @@ import threads.server.domain.like.entity.LikePost;
 import java.util.Optional;
 
 public interface LikePostRepository extends JpaRepository<LikePost, Long>, LikeRepository<LikePost> {
-    Integer countByPostId(Long postId);
+    Long countByPostId(Long postId);
 
     Optional<LikePost> findByUserIdAndPostId(Long userId, Long postId);
 }
